Crate test_listen

Structs

ListenArguments

Functions

bind_fd

Bind the fd to the address.

check_listen_call

get_tests

main

test_after_close

Test listen after closing the socket.

test_backlog_size

Test connecting and accepting sockets with different listen() backlog values.

test_invalid_fd

Test listen using an argument that cannot be a fd.

test_invalid_sock_type

Test listen using an invalid socket type.

test_large_backlog

Test listen using a backlog of INT_MAX.

test_listen_twice

Test calling listen twice for the same socket.

test_listening_not_readable

Test that a listening socket is not readable.

test_listening_not_writable

Test that a listening socket is not writable.

test_negative_backlog

Test listen using a backlog of -1.

test_negative_backlog_connect

Test connecting to a listening socket with a backlog of -1.

test_non_existent_fd

Test listen using an argument that could be a fd, but is not.

test_non_socket_fd

Test listen using a valid fd that is not a socket.

test_reduced_backlog

Test connecting and accepting sockets after the listen() backlog has been lowered.

test_zero_backlog

Test listen using a backlog of 0.