Crate test_listen

Structs

BindAddress

ListenArguments

Functions

bind_fd

Bind the fd to the address.

check_listen_call

get_tests

main

test_after_close

Test listen after closing the socket.

test_invalid_fd

Test listen using an argument that cannot be a fd.

test_invalid_sock_type

Test listen using an invalid socket type.

test_large_backlog

Test listen using a backlog of INT_MAX.

test_listen_twice

Test calling listen twice for the same socket.

test_negative_backlog

Test listen using a backlog of -1.

test_non_existent_fd

Test listen using an argument that could be a fd, but is not.

test_non_socket_fd

Test listen using a valid fd that is not a socket.

test_zero_backlog

Test listen using a backlog of 0.